Function to verify all the sub-dependencies are defined in a list
When working with a list of deps, it's important to know that all the
sub-dependencies have an associated dep_t of the same name in the list.
This means that we can easily find that ordering the deps won't result
stop in an undefined reference.

A new test program has been created for this function.