Return of the strarg functions
After realizing that there would be more uses of dynamic string arrays
than just dep_t, it was decided that the strarg would be resurrected.
The difference now is that there is no strarg structure. Instead, the
caller must supply the necessary types to the functions. This allows
for flexibility in how the strargs are used.