Parse the script line into key and value pairs
Added the static function initd_parse_line to break the lines in the LSB
header into key and value pairs. It just prints these at the end of each
call instead of doing anything useful right now.

There are still issues with the Description field. Some work will have
to be done to handle this correctly.