Check for Required-* keys in initd_list_verify_all()
Changed initd_list_verify_all() to check that all the services in the
Required-Start and Required-Stop fields exist. Added functions
initd_provides and initd_list_provides for checking if a service is
provided by a script.

Added the initd_verify_deps function for checking the fields of a single
script. This is called by initd_verify_all for the Required-* fields.