Set initd_t members from script parser
The parsing code is actually now useful in that it creates an initd_t
structure and sets the members as it reads the supplied script. The
initd_parse function has been changed to take a path and return an
allocated initd_t. The path to the script must be valid. I.e., we no
longer try to guess the directory. This functionality might be added
back later at a higher level.

To support setting the various members, a few functions have been added
to the parsing code. initd_parse_line_tokens takes a line and loops over
the tokens, adding the values found to the provided keys. This will have
to validate input later.

The tparse test has been changed to print the values from the allocated
initd_t.