Skip invalid scripts when parsing
initd_parse() should not return an allocated initd_t if it found that
the script was not valid. Instead, it returns NULL in this case. Callers
of initd_parse() such as initd_list_from_dir() have been converted to
handle the NULL return. Currently, the test of validity is just if the
file has both the opening and closing tags for the LSB header.