dep: Add getter methods for number of deps and dep at index
Add the macros dep_get_dep and dep_get_num for accessing the number of
deps and the dep at a specific index. This will help to hide the
structure of the dep_t.

Converted all the dep_t users except the dep_*() functions to use these
new macros instead of accessing the members directly.