Add dependency recursion function initd_recurse_deps()
Added an initial cut at a dependency recursion and ordering function,
initd_recurse_deps(). This function takes a list of needed deps and
a pool of initd types. From there, it finds each dep in the pool and
recurses its required services for the specified start/stop type.

This function has been added in the new rdep.c file. A test program
has been added to verify functionality.

The implementation needs to be reworked. Due to the use of static dep
types, the function can only be called once.