Create test links at build time and fix for $srcdir != $builddir
The links in the rc?.d directories need to be generated at build time
since we can't distribute symbolic links. Also, if the build directory
is different than the source directory, the target scripts need to be
copied to the build directory so that the links are ../init.d/$script.