Support finding and setting levels for active scripts
By reading the symlinks in the rc*.d directories, the active levels for
init scripts can be set. The current API, initd_list_set_actives(), just
takes an initialized list of initd types and a directory holding them.
This would be /etc/init.d in most cases as it walks from there up and
down through the various rc*.d directories, searching for symlinks to
the init.d directory.

This function needs to be called after parsing the scripts in the init.d
directory. A test program, tactive, has been added to prove out the
functionality. More of the functions in lib/active.c could probably be
made non-static to be more flexible in reading the symlinks, but this is
good enough for now.