Sanitize sysinit handling
This adds more support for the various names that can be used for the
sysinit services. These names the case-insensitive variations of "s",
"si" and "sysinit". This applies to both the rc*.d directory names and
the service levels in the LSB header. The initd_rc_t enum has been
changed to RC_S.

Most of the code is unchanged, but the token-to-rc conversion function
used by the parser has been refactored to handle the sysinit variants
and to be more robust matching against valid levels.