rdep: Append all active services to the requested services
The final order of the dependencies matters when processing recursively
as this will eventually determine the order of the S and K links. Since
this may cause an unchanged link to become out of order when one of its
dependents changes, we add all currently active services to the
requested set. This should ensure that the final order is correct.