Make scripts implicitly provide their names as services
Usually, a script will just provide the services listed in the Provides
field in the LSB header. However, there may be cases where other scripts
depend on it by its name, rather than by the service it provides. This
adds the script's name to its provides member if it is not already
included. This also solves the problem where we want to add a service by
name, but the dep solver operates on provided services.