rdep: Recurse and verify dependencies when removing
Added a new function, initd_remove_recurse_deps, which verifies that the
dependencies of the active services are still valid if specified
services are removed. There is a bit of hackery to play with the active
flags of the to-be-removed services and then restoring them.

A new test program, tremove, has been added to exercise this function.
It is basically the compliment to tverify (should be tinstall).