New initd members: active start and stop symlinks
When the rc?.d directories are walked, we'd like to cache the names of
the links in addition to just flagging the active levels. The names can
be used later when adding or removing links. To this effect, two new
members have been added to initd_t: astart_links and astop_links.