Add functions for installing and removing init.d symlinks
This adds the initd_installrm_links function and its helpers for
installing and removing symlinks. The function takes an ordered start or
kill list and the init.d directory to operate on. There is just a single
entry point since the determination of removal or addition is done in
the dependency ordering.