Use changed fields to fix dep solver in remove case
When told to remove services through initd_remove_recurse_deps, their
changed fields are cleared to indicate that they will no longer be
available. This is an improvement over the previous hack where the
active bits were being cleared and then restored at the end of the
function.

With the changed fields cleared, we can skip services that are currently
active, but will be going away in add_all_actives. This fixes a bug
where a service can be asked to be removed, but then added right back in
because it's currently active.

When each rc level is verified after recursion has completed, the added
services are checked to see if they are marked for removal. This fixes a
bug where a service was asked to be removed, but added back because it
was a dependency of another service. When each service is verified at
that level, it is marked to change to active.

Using the changed field clearing in remove_recurse_deps and the setting
in verify_level, we should be able to install and remove scripts just by
checking the changed field.

Finally, fixed two tests which were passing with false positives due to
these bugs.