Use two fields for storing installation/removal status
Previously, there was a single change field cstart/stop that was
overloaded to contain both installation and removal status. It was
difficult to differentiate between whether a service was to be removed,
to be installed, or unchanged.

This commit converts the change field into two dedicated fields,
instart/stop and rmstart/stop to differentiate the cases better.