Only pass service basenames to the intid functions
The initd functions operate on service names rather than full paths, so
the directory names are stripped off. This is not quite correct as we
really want to just strip the init.d directory name from the beginning
of the path and leave the rest. However, it should work in practice as
only one init.d directory can be used and we expect it to be a flat
directory containing scripts.