Create relative symlinks in /lib/initd
It's nice to create relative symlinks from /lib/initd to /sbin so that
the link path is always correct no matter where the filesystems are
mounted. To accomplish this, a script abs2rel.sh has been added which
converts two directory paths to a relative path. This is used to find
the relative path between /lib/initd and /sbin.