Add automake required files, including GPLv2 license
Added the COPYING, AUTHORS, NEWS and README files. The COPYING file is
standard GPLv2, which we will use for this package. Copyrights will be
added to individual source files later.

The NEWS and README files are currently empty, but will be populated
later. The INSTALL file will also be included in the tarball, but we'll
just use the standard GNU INSTALL file from automake.