Use nodes in list instead of duplicating all initd types
Currently, the entire initd structure is duplicated each time a list is
copied. Since the initd is not specific to each list, this could cause
problems where changing a value in one initd would not be reflected in
the equivalent initd in another list. This also wastes memory as each
initd duplicated when only one copy is needed.

This change introduces the initd_node type. It points to an initd type
and contains the list pointers for inclusion in an initd_list type.
The typical constructors and destructors have been added for handling
these types in lists.

Quite a few places needed to be changed where there were iterations over
lists. We should really add accessor functions and use those whenever
we need access to members of a list.